Se realizo una correcion en el tipo de variables

+++ TvDemo/src/PResoluciones/Resolucion1K.java
@@ -10,8 +10,8 @@ public class Resolucion1K extends Resolucion {
   //
   // Fields
   //
-	protected final static int LARGO = 768;
-	protected final static int ANCHO = 1024;
+	protected final static double LARGO = 768;
+	protected final static double ANCHO = 1024;
 	
 	public Resolucion1K(){
         super(LARGO,ANCHO);
@@ -34,5 +34,6 @@ public class Resolucion1K extends Resolucion {
   //
   // Other methods
   //
+   
 
 }
